[WIP] Package can now be imported and technically run in diffxpy
