+++ batchglm/models/base_glm/__init__.py
@@ -1,4 +1,4 @@
 # from .estimator import _EstimatorGLM
 from ...utils.input import InputDataGLM
 from .model import ModelGLM
-from .utils import closedform_glm_mean, closedform_glm_scale, parse_design
+from .utils import closedform_glm_mean, closedform_glm_scale
