+++ batchglm/utils/__init__.py
@@ -1,2 +1,3 @@
 from . import plotting
 from . import data
+from . import input
