+++ batchglm/utils/input.py
@@ -1,6 +1,5 @@
 import logging
-from operator import indexOf
-from typing import List, Optional, Union
+from typing import List, Optional, Union, Tuple
 
 import anndata
 import dask.array
@@ -10,7 +9,88 @@ import patsy
 import scipy.sparse
 import sparse
 
-from ..models.base_glm.utils import parse_constraints, parse_design
+logger = logging.getLogger("batchglm")
+
+
+def parse_design(
+    design_matrix: Union[pd.DataFrame, patsy.design_info.DesignMatrix, dask.array.core.Array, np.ndarray],
+    param_names: List[str] = None,
+) -> Tuple[np.ndarray, List[str]]:
+    r"""
+    Parser for design matrices.
+
+    :param design_matrix: Design matrix.
+    :param param_names:
+        Optional coefficient names for design_matrix.
+        Ignored if design_matrix is pd.DataFrame or patsy.design_info.DesignMatrix.
+    :return: Tuple[np.ndarray, List[str]] containing the design matrix and the parameter names.
+    :raise AssertionError: if the type of design_matrix is not understood.
+    :raise AssertionError: if length of provided param_names is not equal to number of coefficients in design_matrix.
+    :raise ValueError: if param_names is None when type of design_matrix is numpy.ndarray or dask.array.core.Array.
+    """
+    if isinstance(design_matrix, (pd.DataFrame, patsy.design_info.DesignMatrix)) and param_names is not None:
+        logger.warning(f"The provided param_names are ignored as the design matrix is of type {type(design_matrix)}.")
+
+    if isinstance(design_matrix, patsy.design_info.DesignMatrix):
+        dmat = np.asarray(design_matrix)
+        params = design_matrix.design_info.column_names
+    elif isinstance(design_matrix, pd.DataFrame):
+        dmat = np.asarray(design_matrix)
+        params = design_matrix.columns.tolist()
+    elif isinstance(design_matrix, dask.array.core.Array):
+        dmat = design_matrix.compute()
+        params = param_names
+    elif isinstance(design_matrix, np.ndarray):
+        dmat = design_matrix
+        params = param_names
+    else:
+        raise AssertionError(f"Datatype for design_matrix not understood: {type(design_matrix)}")
+    if params is None:
+        raise ValueError("Provide names when passing design_matrix as np.ndarray or dask.array.core.Array!")
+    assert len(params) == dmat.shape[1], (
+        "Length of provided param_names is not equal to " "number of coefficients in design_matrix."
+    )
+    return dmat, params
+
+
+def parse_constraints(
+    dmat: np.ndarray,
+    dmat_par_names: List[str],
+    constraints: Optional[Union[np.ndarray, dask.array.core.Array]] = None,
+    constraint_par_names: Optional[List[str]] = None,
+) -> Tuple[np.ndarray, List[str]]:
+    r"""
+    Parser for constraint matrices.
+
+    :param dmat: Design matrix.
+    :param constraints: Constraint matrix.
+    :param constraint_par_names: Optional coefficient names for constraints.
+    :return: Tuple[np.ndarray, List[str]] containing the constraint matrix and the parameter names.
+    :raise AssertionError: if the type of given design / contraint matrix is not np.ndarray or dask.array.core.Array.
+    """
+    assert isinstance(dmat, np.ndarray), "dmat must be provided as np.ndarray."
+    if constraints is None:
+        constraints = np.identity(n=dmat.shape[1])
+        constraint_params = dmat_par_names
+    else:
+        if isinstance(constraints, dask.array.core.Array):
+            constraints = constraints.compute()
+        assert isinstance(constraints, np.ndarray), "contraints must be np.ndarray or dask.array.core.Array."
+        # Cannot use all parameter names if constraint matrix is not identity: Make up new ones.
+        # Use variable names that can be mapped (unconstrained).
+        if constraint_par_names is not None:
+            assert len(constraint_params) == len(constraint_par_names)
+            constraint_params = constraint_par_names
+        else:
+            constraint_params = [
+                "var_" + str(i)
+                if np.sum(constraints[:, i] != 0) > 1
+                else dmat_par_names[np.where(constraints[:, i] != 0)[0][0]]
+                for i in range(constraints.shape[1])
+            ]
+        assert constraints.shape[0] == dmat.shape[1], "constraint dimension mismatch"
+
+    return constraints, constraint_params
 
 
 class InputDataGLM:
